解决三级菜单信息错乱问题

+++ app/src/main/java/com/qhn/bhne/footprinting/interfaces/PopClickItemCallBack.java
@@ -8,6 +8,6 @@ import android.view.MenuItem;
  */
 
 public interface PopClickItemCallBack {
-    void clickItem(MenuItem menuItem,int createCategory,int projectID);
+    void clickItem(MenuItem menuItem,int createCategory,long projectID);
     //void clickListItem(MenuItem menuItem,int createCategory,int position);
 }
